AgentRetrival/requirements.txt
accelerate>=0.33.0
bitsandbytes>=0.43.1 ; platform_system != 'Windows'
transformers>=4.43.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
torch>=2.2.0
numpy>=1.26
pydantic>=2.7
fastapi>=0.111
uvicorn>=0.30
pandas>=2.2
python-docx>=1.1
pymupdf>=1.24.4
scikit-learn>=1.4
nltk>=3.9
regex>=2024.5.15
rapidfuzz>=3.9
orjson>=3.10
# optionnels (modèles locaux de modération)
detoxify>=0.5.2 ; extra == 'moderation'